Rename benchmarks to mirror astropy core package structure
