+++ benchmarks/io_ascii/__init__.py
@@ -1,10 +1,10 @@
 __all__ = [
-    "time_ascii",
-    "time_comparison",
-    "time_core",
-    "time_fixedwidth",
-    "time_ipac",
-    "time_rdb",
-    "time_sextractor",
-    "time_table"
+    "ascii",
+    "comparison",
+    "core",
+    "fixedwidth",
+    "ipac",
+    "rdb",
+    "sextractor",
+    "table"
 ]
