+++ benchmarks/io_ascii/comparison.py
