+++ benchmarks/io_ascii/rdb.py
